+++ src/app/features/cart/cart.module.ts
@@ -1,5 +1,4 @@
 import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 import { CartListComponent } from './components/cart-list/cart-list.component';
 import { CartItemComponent } from './components/cart-item/cart-item.component';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
@@ -18,7 +17,6 @@ const LOCAL_COMPONENTS = [
     ...LOCAL_COMPONENTS
   ],
   imports: [
-    CommonModule,
     SharedModule,
     FontAwesomeModule
   ],